project file saving and geometry
files now open with proper canvas and dialog size
files are now saved in new format, albeit with concentrated
code that gathers all of the information (as opposed to the
.xml() methods used previously).
I'm not sure I want to change that.  As a matter of fact,
I want all of the objects to just use the project file
so it's always up to date.